MAINT: Fix Cython/C type problems
- Single-source Cython/C type definitions: `core._ext.types`
- Introduce semantic type aliases: ADJ, NODE, DEGREE, WEIGHT, FIELD, etc.
- Remove deprecated type aliases: `np.{int|float}`
- Replace C type aliases by Numpy type aliases
- Add missing Numpy type casts before Python/Cython boundary
- Disambiguate `timeseries._ext.numerics._twin_surrogates_{s|r}`
  (blame: 098489b & 2f6a5b9)
- Fix some broken output pointers
- Reduce redundant copying
- Vectorize some simple Python loops
- Eliminate some dead code
- Potentially related issues: #128, #141, #142, #145

+++ pyunicorn/climate/_ext/src_numerics.c
@@ -8,6 +8,8 @@
 * License: BSD (3-clause)
 */
 
+#include <math.h>
+
 // mutual_info ================================================================
 
 void _cython_calculate_mutual_information(float *anomaly, int n_samples,
@@ -18,7 +20,7 @@ void _cython_calculate_mutual_information(float *anomaly, int n_samples,
          in_nodes;
     double norm, rescaled, hpl, hpm, plm;
 
-    double *p_anomaly;
+    float *p_anomaly;
     float *p_mi, *p_mi2;
     long *p_symbolic, *p_symbolic1, *p_symbolic2, *p_hist, *p_hist1,
          *p_hist2, *p_hist2d;
